Rename panel module
