+++ src/fontra/views/editor/panel-related-glyphs.js
